+++ test/sync/src/commonMain/kotlin/io/realm/test/mongodb/TestApp.kt
@@ -18,7 +18,7 @@
 
 package io.realm.test.mongodb
 
-import io.ktor.client.request.get
+import io.ktor.client.request.*
 import io.realm.internal.platform.runBlocking
 import io.realm.internal.platform.singleThreadDispatcher
 import io.realm.mongodb.App
